+++ R/roxygen-templates.R
@@ -25,7 +25,7 @@
 # - `PREFIX_auth_configure_params() crosslinks to
 #   `gargle::gargle_oauth_client_from_json()` which requires gargle (>= 1.3.0)
 
-glue_data_lines <- function(.data, lines, ..., .envir = parent.frame()) {
+glue_data_lines <- function(.data, lines, ..., .envir = caller_env()) {
   # work around name collision of `.x` of map_chr() vs. of glue_data()
   # and confusion re: `...` of glue_data_lines() vs. `...` of map_chr()
   # plus: I've only got compat-purrr here, so I have to write a function
@@ -51,14 +51,7 @@ PREFIX_auth_description <- function(.data = list(
     "credentials are cached in a folder below your home directory, from where",
     "they can be automatically refreshed, as necessary. Storage at the user",
     "level means the same token can be used across multiple projects and",
-    "tokens are less likely to be synced to the cloud by accident.",
-    "",
-    "If you are interacting with R within a browser (applies to RStudio Server,",
-    "Posit Workbench, and Posit Cloud), you need a variant of this flow,",
-    "known as out-of-band auth (\"oob\"). If this does not happen",
-    "automatically, you can request it yourself with `use_oob = TRUE` or,",
-    "more persistently, by setting an option via",
-    "`options(gargle_oob_default = TRUE)`."
+    "tokens are less likely to be synced to the cloud by accident."
   ), .data = .data)
 }
 
@@ -71,26 +64,40 @@ PREFIX_auth_details <- function(.data = list(
     "Most users, most of the time, do not need to call `{PREFIX}_auth()`",
     "explicitly -- it is triggered by the first action that requires",
     "authorization. Even when called, the default arguments often suffice.",
-    "However, when necessary, this function allows the user to explicitly:",
-    "  * Declare which Google identity to use, via an email address. If there",
-    "    are multiple cached tokens, this can clarify which one to use. It can",
-    "    also force {PACKAGE} to switch from one identity to another. If",
-    "    there's no cached token for the email, this triggers a return to the",
-    "    browser to choose the identity and give consent. You can specify just",
-    "    the domain by using a glob pattern. This means that a script",
-    "    containing `email = \"*@example.com\"` can be run without further",
-    "    tweaks on the machine of either `alice@example.com` or",
-    "    `bob@example.com`.",
-    "  * Use a service account token or workload identity federation.",
-    "  * Bring their own [Token2.0][httr::Token-class].",
-    "  * Specify non-default behavior re: token caching and out-of-bound",
-    "    authentication.",
-    "  * Customize scopes.",
+    "",
+    "However, when necessary, `{PREFIX}_auth()` allows the user to explicitly:",
+    "  * Declare which Google identity to use, via an `email` specification.",
+    "  * Use a service account token or workload identity federation via",
+    "    `path`.",
+    "  * Bring your own `token`.",
+    "  * Customize `scopes`.",
+    "  * Use a non-default `cache` folder or turn caching off.",
+    "  * Explicitly request out-of-bound (OOB) auth via `use_oob`.",
+    "",
+    "If you are interacting with R within a browser (applies to RStudio",
+    "Server, Posit Workbench, Posit Cloud, and Google Colaboratory), you need",
+    "OOB auth or the pseudo-OOB variant. If this does not happen",
+    "automatically, you can request it explicitly with `use_oob = TRUE` or,",
+    "more persistently, by setting an option via",
+    "`options(gargle_oob_default = TRUE)`.",
+    "",
+    "The choice between conventional OOB or pseudo-OOB auth is determined",
+    "by the type of OAuth client. If the client is of the \"installed\" type,",
+    "`use_oob = TRUE` results in conventional OOB auth. If the client is of",
+    "the \"web\" type, `use_oob = TRUE` results in pseudo-OOB auth. Packages",
+    "that provide a built-in OAuth client can usually detect which type of",
+    "client to use. But if you need to set this explicitly, use the",
+    "`\"gargle_oauth_client_type\"` option:",
+    "```r",
+    "options(gargle_oauth_client_type = \"web\")       # pseudo-OOB",
+    "# or, alternatively",
+    "options(gargle_oauth_client_type = \"installed\") # conventional OOB",
+    "```",
     "",
     "For details on the many ways to find a token, see",
     "[gargle::token_fetch()]. For deeper control over auth, use",
     "[{PREFIX}_auth_configure()] to bring your own OAuth client or API key.",
-    "Read more about gargle options, see [gargle::gargle_options]."
+    "To learn more about gargle options, see [gargle::gargle_options]."
   ), .data = .data)
 }